[Feat] 데미지 자료형 int -> flaot 변경

+++ Assets/Workers/DEV/YTH/Scripts/Monster/PlayerControllerMonster.cs
@@ -4,9 +4,9 @@ using UnityEngine;
 
 public class PlayerControllerMonster : MonoBehaviour, IDamagable
 {
-    [SerializeField] int curHp;
-    public int CurHp { get { return curHp; } set { curHp = value; } }
-    [SerializeField] int maxHp = 100;
+    [SerializeField] float curHp;
+    public float CurHp { get { return curHp; } set { curHp = value; } }
+    [SerializeField] float maxHp = 100;
 
     private void Start()
     {
